NHD Flowline (ComID) Attributes

NHD Flowline (ComID) Attributes

flowline_attr

Format

An object of class tbl_df (inherits from tbl, data.frame) with 178868 rows and 162 columns.

Examples

flowline_attr |> pillar::glimpse()
#> Rows: 178,868
#> Columns: 162
#> $ comid                        <dbl> 20245062, 24085230, 22226684, 22226720, 22226732, 22226762, 22226782, 22226810, 22226834, 22226848, 22226864, 22226866, 22226870…
#> $ reachcode                    <fct> 16060014044574, 17120005008721, 18010101000020, 18010101000021, 18010101000023, 18010101000024, 18010101000025, 18010101000026, …
#> $ gnis_id                      <chr> NA, NA, "253899", "253899", "253899", "253899", "253899", "253899", "263616", "263616", "263616", "263616", "263616", "263616", …
#> $ gnis_name                    <chr> NA, NA, "Smith River", "Smith River", "Smith River", "Smith River", "Smith River", "Smith River", "Mill Creek", "Mill Creek", "M…
#> $ lengthkm                     <dbl> 5.415, 2.787, 0.598, 2.855, 0.807, 5.394, 4.668, 2.622, 3.486, 1.599, 1.215, 0.292, 0.422, 2.608, 1.053, 0.656, 1.795, 0.811, 3.…
#> $ ftype                        <fct> StreamRiver, StreamRiver, StreamRiver, StreamRiver, StreamRiver, StreamRiver, StreamRiver, StreamRiver, StreamRiver, StreamRiver…
#> $ fcode                        <fct> 46003, 46003, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46006, 46…
#> $ rc_huc_8                     <chr> "16060014", "17120005", "18010101", "18010101", "18010101", "18010101", "18010101", "18010101", "18010101", "18010101", "1801010…
#> $ rc_huc_10                    <chr> "1606001404", "1712000500", "1801010100", "1801010100", "1801010100", "1801010100", "1801010100", "1801010100", "1801010100", "1…
#> $ rc_huc_12                    <chr> "160600140445", "171200050087", "180101010000", "180101010000", "180101010000", "180101010000", "180101010000", "180101010000", …
#> $ ftype_desc                   <chr> "Stream/River", "Stream/River", "Stream/River", "Stream/River", "Stream/River", "Stream/River", "Stream/River", "Stream/River", …
#> $ artificial_path              <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ ply_comid                    <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ ply_gnis_id                  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ ply_gnis_name                <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ huc_8                        <chr> NA, NA, "18010101", "18010101", "18010101", "18010101", "18010101", "18010101", "18010101", "18010101", "18010101", "18010101", …
#> $ huc_10                       <chr> NA, NA, "1801010104", "1801010104", "1801010104", "1801010104", "1801010104", "1801010104", "1801010104", "1801010104", "1801010…
#> $ hu_10_name                   <chr> NA, NA, "Smith River", "Smith River", "Smith River", "Smith River", "Smith River", "Smith River", "Smith River", "Smith River", …
#> $ huc_12                       <chr> NA, NA, "180101010404", "180101010404", "180101010404", "180101010404", "180101010404", "180101010404", "180101010402", "1801010…
#> $ hu_12_type                   <chr> NA, NA, "M", "M", "M", "M", "M", "M", "S", "S", "S", "S", "S", "S", "S", "M", "M", "M", "S", "S", "S", "S", "S", "S", "S", "S", …
#> $ hu_12_name                   <chr> NA, NA, "Smith River", "Smith River", "Smith River", "Smith River", "Smith River", "Smith River", "Mill Creek", "Mill Creek", "M…
#> $ hu_12_ds                     <chr> NA, NA, "180101010502", "180101010502", "180101010502", "180101010502", "180101010502", "180101010502", "180101010404", "1801010…
#> $ hydro_seq                    <dbl> NA, NA, 10013332, 10013490, 10013972, 10014142, 10014325, 10014511, 10030080, 10031409, 10032891, 10034617, 10036591, 10038903, …
#> $ reach_code                   <fct> NA, NA, 18010101000020, 18010101000021, 18010101000023, 18010101000024, 18010101000025, 18010101000026, 18010101000027, 18010101…
#> $ stream_level                 <dbl> NA, NA, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 3, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 2, …
#> $ stream_order                 <dbl> NA, NA, 5, 5, 5, 5, 5, 5, 4, 4, 4, 4, 4, 4, 3, 5, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 3, 3, 3, 3, 3, 2, 1, 4, …
#> $ us_length_km                 <dbl> NA, NA, 1325.477, 1316.741, 1310.090, 1304.299, 1292.314, 1283.472, 84.641, 79.109, 75.340, 72.107, 69.757, 63.655, 23.106, 1196…
#> $ ds_length_km                 <dbl> NA, NA, 0.000, 0.598, 4.892, 5.699, 11.093, 15.761, 18.383, 21.869, 23.468, 24.683, 24.975, 25.397, 28.005, 18.383, 19.039, 20.8…
#> $ da_area_sq_km                <dbl> NA, NA, 1755.1854, 1746.2691, 1738.4283, 1728.0855, 1705.6791, 1689.4917, 95.7528, 89.2260, 84.4875, 81.1359, 79.6392, 74.9817, …
#> $ da_area_sq_km_div            <dbl> NA, NA, 1755.1854, 1746.2691, 1738.4283, 1728.0855, 1705.6791, 1689.4917, 95.7528, 89.2260, 84.4875, 81.1359, 79.6392, 74.9817, …
#> $ da_area_sq_km_tot            <dbl> NA, NA, 1755.1854, 1746.2691, 1738.4283, 1728.0855, 1705.6791, 1689.4917, 95.7528, 89.2260, 84.4875, 81.1359, 79.6392, 74.9817, …
#> $ divergence_ratio             <dbl> NA, NA, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ reach_length_km              <dbl> NA, NA, 0.598, 2.855, 0.807, 5.394, 4.668, 2.622, 3.486, 1.599, 1.215, 0.292, 0.422, 2.608, 1.053, 0.656, 1.795, 0.811, 3.351, 2…
#> $ reach_length_ft              <dbl> NA, NA, 1961.94226, 9366.79790, 2647.63780, 17696.85039, 15314.96063, 8602.36220, 11437.00787, 5246.06299, 3986.22047, 958.00525…
#> $ slope                        <dbl> NA, NA, 0.00065217, 0.00069001, 0.00003717, 0.00081757, 0.00190017, 0.00133485, 0.00570281, 0.00213258, 0.00795884, 0.00339041, …
#> $ elev_min                     <dbl> NA, NA, 3.05, 3.44, 8.69, 8.72, 13.13, 22.00, 25.50, 45.38, 48.79, 58.46, 59.45, 63.61, 68.73, 25.50, 27.20, 46.08, 46.08, 65.32…
#> $ elev_max                     <dbl> NA, NA, 3.44, 5.41, 8.72, 13.13, 22.00, 25.50, 45.38, 48.79, 58.46, 59.45, 63.61, 68.73, 73.33, 27.20, 46.08, 46.08, 65.32, 73.2…
#> $ stream_power                 <dbl> NA, NA, 1.144679262, 1.204943142, 0.064617380, 1.412830862, 3.241080255, 2.255217996, 0.546060025, 0.190281583, 0.672422495, 0.2…
#> $ da_ppt_mean_mm               <dbl> NA, NA, 2723.898, 2727.656, 2731.168, 2733.134, 2740.761, 2746.749, 2351.744, 2364.892, 2381.114, 2388.872, 2391.910, 2408.075, …
#> $ loc_ppt_mean_mm              <dbl> NA, NA, 1821.85, 1910.73, 1939.38, 2097.04, 2040.72, 2119.72, 2146.45, 2115.43, 2108.63, 2130.96, 2130.50, 2168.98, 2165.49, 212…
#> $ vogel_q_ma_cfs               <dbl> NA, NA, 6178.05072, 6165.75802, 6155.65922, 6130.26554, 6090.33734, 6063.74845, 259.42945, 245.27943, 236.04806, 228.56757, 225.…
#> $ vogel_v_ma_fps               <dbl> NA, NA, 1.81167, 1.82382, 1.26003, 1.86210, 2.08068, 1.98094, 1.37341, 1.20561, 1.40226, 1.25526, 1.42591, 1.16226, 1.08374, 2.1…
#> $ erom_q_ma_cfs                <dbl> NA, NA, 4127.598, 4111.549, 4097.949, 4074.883, 4031.351, 3999.819, 142.626, 133.011, 126.478, 121.548, 119.334, 112.727, 37.446…
#> $ erom_v_ma_fps                <dbl> NA, NA, NA, NA, 1.22917, 1.81155, 2.02242, 1.92551, 1.32303, 1.16222, 1.34835, 1.20805, 1.36991, 1.11861, 1.03836, 2.08110, 2.53…
#> $ sinuosity                    <dbl> 1.07, 1.69, 1.04, 1.10, 1.11, 1.13, 1.26, 1.30, 1.32, 1.55, 1.22, 1.03, 1.09, 1.24, 1.18, 1.13, 1.05, 1.01, 1.27, 1.19, 1.38, 1.…
#> $ da_avg_slope                 <dbl> NA, NA, 42.38, 42.51, 42.65, 42.75, 42.99, 43.23, 30.59, 30.93, 31.80, 32.15, 32.25, 32.79, 32.82, 44.03, 44.11, 44.20, 45.91, 4…
#> $ da_elev_mean                 <dbl> NA, NA, 717.08, 720.02, 722.78, 725.09, 732.31, 737.95, 279.17, 288.51, 298.19, 303.52, 305.74, 312.87, 277.57, 766.96, 768.83, …
#> $ da_elev_min                  <dbl> NA, NA, 3.05, 3.30, 5.54, 5.54, 12.59, 19.71, 25.50, 43.88, 48.50, 58.40, 58.75, 59.76, 68.72, 25.33, 26.60, 30.63, 34.05, 58.63…
#> $ da_elev_max                  <dbl> NA, NA, 1927.99, 1927.99, 1927.99, 1927.99, 1927.99, 1927.99, 715.50, 715.50, 715.50, 715.50, 715.50, 715.50, 660.08, 1927.99, 1…
#> $ da_elev_rel                  <dbl> NA, NA, 1924.94, 1924.69, 1922.45, 1922.45, 1915.40, 1908.28, 690.00, 671.62, 667.00, 657.10, 656.75, 655.74, 591.36, 1902.66, 1…
#> $ loc_bfi                      <dbl> NA, NA, 46.0000, 45.9753, 45.0000, 44.7088, 44.0000, 44.0000, 44.0000, 44.0000, 44.0000, 44.0000, 44.0000, 44.0000, 44.0196, 44.…
#> $ da_bfi                       <dbl> NA, NA, 46.1662, 46.1671, 46.1697, 46.1678, 46.1893, 46.2102, 44.5755, 44.6176, 44.6522, 44.6792, 44.6919, 44.7349, 44.8440, 46.…
#> $ loc_twi                      <dbl> NA, NA, 986.5119, 910.0642, 690.3582, 704.8816, 747.9005, 710.4250, 666.2432, 733.8008, 711.1087, 780.2000, 743.9758, 685.5036, …
#> $ da_twi                       <dbl> NA, NA, 602.5557, 601.6043, 600.5490, 600.2391, 599.0380, 597.8346, 628.3256, 626.3497, 621.5346, 619.4696, 619.0707, 617.7287, …
#> $ loc_pct_clay                 <dbl> NA, NA, 23.5595, 25.7558, 33.3100, 31.2657, 29.2201, 32.8140, 34.7125, 34.7300, 34.7300, 34.7300, 34.7300, 34.7300, 34.7300, 30.…
#> $ da_pct_clay                  <dbl> NA, NA, 26.7798, 26.7672, 26.7629, 26.7276, 26.6721, 26.6498, 34.4717, 34.4538, 34.4383, 34.4262, 34.4205, 34.4013, 34.7300, 26.…
#> $ loc_pct_sand                 <dbl> NA, NA, 29.0960, 25.0422, 14.7206, 18.2691, 22.7085, 21.7815, 19.1185, 19.0000, 19.0000, 19.0000, 19.0000, 19.0000, 19.0000, 24.…
#> $ da_pct_sand                  <dbl> NA, NA, 29.0673, 29.1089, 29.1428, 29.2196, 29.3523, 29.4076, 19.3687, 19.3890, 19.4108, 19.4278, 19.4358, 19.4629, 19.0000, 30.…
#> $ loc_permeability             <dbl> NA, NA, 10.9010, 7.9839, 1.9500, 3.0299, 4.0528, 2.0863, 1.9413, 1.9400, 1.9400, 1.9400, 1.9400, 1.9400, 1.9400, 2.2500, 2.2485,…
#> $ loc_bedrock_depth            <dbl> NA, NA, 148.5053, 137.5721, 110.2300, 120.1530, 136.7819, 115.8794, 139.0963, 139.3100, 139.3100, 139.3100, 139.3100, 139.3100, …
#> $ da_permeability              <dbl> NA, NA, 3.1762, 3.1594, 3.1475, 3.1542, 3.1563, 3.1507, 1.9592, 1.9605, 1.9617, 1.9626, 1.9630, 1.9644, 1.9400, 3.2248, 3.2290, …
#> $ da_bedrock_depth             <dbl> NA, NA, 102.8035, 102.6739, 102.5554, 102.5356, 102.3613, 102.1049, 136.7739, 136.6005, 136.4485, 136.3303, 136.2743, 136.0857, …
#> $ loc_k_erodibility            <dbl> NA, NA, 0.3163, 0.3113, 0.3094, 0.3201, 0.3540, 0.3794, 0.3900, 0.3900, 0.3900, 0.3900, 0.3900, 0.3900, 0.3900, 0.3875, 0.3809, …
#> $ da_k_erodibility             <dbl> NA, NA, 0.3540, 0.3542, 0.3544, 0.3547, 0.3552, 0.3551, 0.3895, 0.3894, 0.3894, 0.3894, 0.3894, 0.3893, 0.3900, 0.3530, 0.3529, …
#> $ loc_precip                   <dbl> NA, NA, 1793.005, 1863.625, 1877.664, 2014.618, 1949.340, 2003.199, 2016.543, 1998.174, 1996.843, 2018.341, 2010.967, 2039.757, …
#> $ da_precip                    <dbl> NA, NA, 2696.198, 2700.126, 2703.767, 2706.316, 2714.807, 2721.530, 2226.069, 2239.800, 2254.915, 2262.858, 2266.169, 2282.424, …
#> $ loc_runoff                   <dbl> NA, NA, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, …
#> $ da_runoff                    <dbl> NA, NA, 2142.698, 2142.661, 2142.628, 2142.583, 2142.486, 2142.414, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, 2150.000, …
#> $ bf_width_m                   <dbl> NA, NA, 54.37433, 54.26395, 58.87000, 58.50000, 59.02000, 58.82000, 19.73000, 19.66000, 18.76000, 19.20000, 18.62000, 19.04000, …
#> $ bf_depth_m                   <dbl> NA, NA, 2.068039, 2.064945, 2.790000, 2.300000, 2.300000, 2.290000, 1.470000, 1.490000, 1.400000, 1.440000, 1.400000, 1.470000, …
#> $ bf_w_d_ratio                 <dbl> NA, NA, 26.292698, 26.278642, 21.100358, 25.434783, 25.660870, 25.685590, 13.421769, 13.194631, 13.400000, 13.333333, 13.300000,…
#> $ mean_ndvi                    <dbl> NA, NA, 0.6854199, 0.7078681, 0.6419506, 0.7930269, 0.7990128, 0.7999287, 0.8299923, 0.8251462, 0.8445378, 0.8372284, 0.8452153,…
#> $ bio_aq_rank_sw               <int> NA, NA, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, …
#> $ species                      <dbl> NA, NA, 281, 281, 281, 281, 281, 281, 139, 139, 139, 139, 139, 139, 139, 281, 281, 281, 66, 66, 66, 66, 66, 66, 66, 66, 66, 66, …
#> $ species_fish                 <dbl> NA, NA, 16, 16, 16, 16, 16, 16, 11, 11, 11, 11, 11, 11, 11, 16, 16, 16, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, …
#> $ species_crust                <dbl> NA, NA, 3, 3, 3, 3, 3, 3, 0, 0, 0, 0, 0, 0, 0, 3, 3, 3, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ species_herps                <dbl> NA, NA, 16, 16, 16, 16, 16, 16, 14, 14, 14, 14, 14, 14, 14, 16, 16, 16, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, …
#> $ species_inverts              <dbl> NA, NA, 136, 136, 136, 136, 136, 136, 70, 70, 70, 70, 70, 70, 70, 136, 136, 136, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ species_mollusks             <dbl> NA, NA, 6, 6, 6, 6, 6, 6, 2, 2, 2, 2, 2, 2, 2, 6, 6, 6, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, …
#> $ species_plants               <dbl> NA, NA, 37, 37, 37, 37, 37, 37, 24, 24, 24, 24, 24, 24, 24, 37, 37, 37, 31, 31, 31, 31, 31, 31, 31, 31, 31, 31, 31, 31, 31, 31, …
#> $ species_birds                <dbl> NA, NA, 62, 62, 62, 62, 62, 62, 13, 13, 13, 13, 13, 13, 13, 62, 62, 62, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 1, 1, 1,…
#> $ species_mammals              <dbl> NA, NA, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, …
#> $ species_mollusks_crust       <dbl> NA, NA, 9, 9, 9, 9, 9, 9, 2, 2, 2, 2, 2, 2, 2, 9, 9, 9, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, …
#> $ species_endemic              <dbl> NA, NA, 32, 32, 32, 32, 32, 32, 16, 16, 16, 16, 16, 16, 16, 32, 32, 32, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 1, 1, 1,…
#> $ species_endemic_fish         <dbl> NA, NA, 3, 3, 3, 3, 3, 3, 1, 1, 1, 1, 1, 1, 1, 3, 3, 3, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 1, 1, 1, 1, 1, 1, 1, 2, …
#> $ species_endemic_crust        <dbl> NA, NA, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ species_endemic_herps        <dbl> NA, NA, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ species_endemic_inverts      <dbl> NA, NA, 26, 26, 26, 26, 26, 26, 13, 13, 13, 13, 13, 13, 13, 26, 26, 26, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ species_endemic_mollusks     <dbl> NA, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ species_endemic_plants       <dbl> NA, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 0, 0, 0, 0, 0, 0, 0, 2, …
#> $ species_endemic_birds        <dbl> NA, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ species_endemic_mammals      <dbl> NA, NA, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ species_vulnerable           <dbl> NA, NA, 66, 66, 66, 66, 66, 66, 36, 36, 36, 36, 36, 36, 36, 66, 66, 66, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, …
#> $ species_listed               <dbl> NA, NA, 6, 6, 6, 6, 6, 6, 2, 2, 2, 2, 2, 2, 2, 6, 6, 6, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ species_endemic_vulnerable   <dbl> NA, NA, 11, 11, 11, 11, 11, 11, 4, 4, 4, 4, 4, 4, 4, 11, 11, 11, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, …
#> $ species_endemic_listed       <dbl> NA, NA, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ genus                        <dbl> NA, NA, 224, 224, 224, 224, 224, 224, 120, 120, 120, 120, 120, 120, 120, 224, 224, 224, 54, 54, 54, 54, 54, 54, 54, 54, 54, 54, …
#> $ family                       <dbl> NA, NA, 111, 111, 111, 111, 111, 111, 76, 76, 76, 76, 76, 76, 76, 111, 111, 111, 44, 44, 44, 44, 44, 44, 44, 44, 44, 44, 44, 44,…
#> $ tax_order                    <dbl> NA, NA, 56, 56, 56, 56, 56, 56, 39, 39, 39, 39, 39, 39, 39, 56, 56, 56, 33, 33, 33, 33, 33, 33, 33, 33, 33, 33, 33, 33, 33, 33, …
#> $ tax_class                    <dbl> NA, NA, 15, 15, 15, 15, 15, 15, 13, 13, 13, 13, 13, 13, 13, 15, 15, 15, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, 11, …
#> $ phylum                       <dbl> NA, NA, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 3, 3, 3, 3, 3, 3, 3, 4, …
#> $ species_current              <dbl> NA, NA, 227, 227, 227, 227, 227, 227, 110, 110, 110, 110, 110, 110, 110, 227, 227, 227, 22, 22, 22, 22, 22, 22, 22, 22, 22, 22, …
#> $ species_historical           <dbl> NA, NA, 64, 64, 64, 64, 64, 64, 30, 30, 30, 30, 30, 30, 30, 64, 64, 64, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, 39, …
#> $ species_other                <dbl> NA, NA, 36, 36, 36, 36, 36, 36, 29, 29, 29, 29, 29, 29, 29, 36, 36, 36, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, 28, …
#> $ hyd_cat                      <fct> NA, NA, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, Rain, …
#> $ hyd_cls                      <fct> NA, NA, "Winter storms", "Winter storms", "Winter storms", "Winter storms", "Winter storms", "Winter storms", "Winter storms", "…
#> $ nf_bfl_dry_cfs_dry           <dbl> NA, NA, 350.00, 348.00, 347.00, 344.00, 340.00, 337.00, 8.73, 8.21, 7.77, 7.51, 7.37, 6.94, 2.41, 317.00, 316.00, 315.00, 151.00…
#> $ nf_bfl_dry_cfs               <dbl> NA, NA, 357.00, 355.00, 353.00, 351.00, 347.00, 344.00, 11.70, 11.10, 10.60, 10.10, 9.96, 9.39, 3.12, 323.00, 322.00, 321.00, 15…
#> $ nf_bfl_wet_cfs               <dbl> NA, NA, 1790.0, 1780.0, 1770.0, 1760.0, 1750.0, 1730.0, 64.2, 60.2, 57.6, 55.6, 54.7, 52.2, 19.4, 1640.0, 1630.0, 1630.0, 785.0,…
#> $ nf_wet_start                 <dbl> NA, NA, 37.0, 37.0, 37.0, 37.0, 37.0, 38.0, 40.3, 40.0, 40.0, 39.9, 39.8, 39.8, 42.8, 38.0, 38.0, 38.0, 37.0, 37.0, 37.0, 37.0, …
#> $ nf_wet_dur                   <dbl> NA, NA, 167, 167, 167, 166, 166, 166, 172, 171, 171, 171, 171, 171, 171, 166, 166, 166, 167, 167, 167, 167, 167, NA, 166, 166, 1…
#> $ peak_q2_cfs                  <dbl> NA, NA, 60800, 60500, 60200, 59900, 59100, 58500, 2780, 2590, 2450, 2360, 2310, 2180, 837, 55100, 54900, 54700, 26900, 25100, 24…
#> $ peak_q5_cfs                  <dbl> NA, NA, 84300, 83900, 83500, 83000, 82000, 81200, 4110, 3830, 3620, 3480, 3410, 3210, 1240, 76400, 76200, 75800, 30400, 28400, 2…
#> $ peak_q10_cfs                 <dbl> NA, NA, 120000, 119000, 119000, 118000, 116000, 115000, 4460, 4150, 4110, 3950, 3860, 3890, 1760, 108000, 108000, 108000, 39300,…
#> $ merit_width_m                <dbl> NA, NA, NA, 70.40650, 33.88020, 53.46577, 41.81242, 44.81628, 44.10711, NA, NA, NA, NA, NA, NA, 22.23854, 37.86823, 43.62812, 22…
#> $ vb_width_transect            <dbl> NA, NA, 54.37433, 70.40650, 33.88020, 53.46577, 41.81242, 44.81628, 44.10711, 19.66000, 18.76000, 19.20000, 18.62000, 19.04000, …
#> $ frac_leveed_longitudinal     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ lateral_levee_confinement_ft <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ geomorph_class               <fct> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ geomorph_confined            <fct> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ geomorph_uniform             <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ geomorph_steppool            <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ geomorph_riffles             <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ geomorph_gravel              <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ geomorph_spawning            <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ hqt_gradient_class           <fct> Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Bedrock, Be…
#> $ range_FR_Extant              <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_FR_Historical          <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_LFR_Extant             <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_LFR_Historical         <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_SR_Extant              <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_SR_Historical          <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_SR_Observed            <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_WR_Extant              <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_WR_Historical          <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_WR_Observed            <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_cvchinook_historical   <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ range_cvchinook_extant       <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FA…
#> $ river_group                  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ river_cvpia                  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ habitat                      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ level                        <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ watershed_cvpia_parent       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ watershed_cvpia              <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ watershed_level_1            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ watershed_level_2            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ watershed_level_3            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ range_pisces                 <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ bf_xarea_m                   <dbl> NA, NA, 112.44826, 112.05209, 164.24730, 134.55000, 135.74600, 134.69780, 29.00310, 29.29340, 26.26400, 27.64800, 26.06800, 27.9…
#> $ chan_width_m                 <dbl> NA, NA, 54.37433, 70.40650, 33.88020, 53.46577, 41.81242, 44.81628, 44.10711, 19.66000, 18.76000, 19.20000, 18.62000, 19.04000, …
#> $ velocity_m_s                 <dbl> NA, NA, NA, NA, 4.032710, 5.943406, 6.635236, 6.317290, 4.340650, 3.813058, 4.423720, 3.963419, 4.494455, 3.669980, 3.406693, 6.…
#> $ wetted_perimeter_m           <dbl> NA, NA, 58.51041, 58.39384, 64.45000, 63.10000, 63.62000, 63.40000, 22.67000, 22.64000, 21.56000, 22.08000, 21.42000, 21.98000, …
#> $ hydraulic_radius_m           <dbl> NA, NA, 1.9218504, 1.9189024, 2.5484453, 2.1323296, 2.1337001, 2.1245710, 1.2793604, 1.2938781, 1.2181818, 1.2521739, 1.2169935,…
#> $ critical_shields_number      <dbl> NA, NA, 0.02397074, 0.02431113, 0.01171222, 0.02536428, 0.03131766, 0.02867143, 0.04122052, 0.03223425, 0.04480265, 0.03619548, …
#> $ grain_size_mobilized_mm      <dbl> NA, NA, 0.31689472, 0.33008001, 0.04901675, 0.41655548, 0.78460761, 0.59947405, 1.07271505, 0.51879664, 1.31151795, 0.71085106, …
#> $ shear_velocity_cm_s          <dbl> NA, NA, 11.088549, 11.396950, 3.048375, 13.077483, 19.943318, 16.679628, 26.753180, 16.452574, 30.840077, 20.407647, 34.305902, …
#> $ settling_velocity_ndim       <dbl> NA, NA, 8.423085e+01, 9.145616e+01, 1.750058e+00, 1.381718e+02, 4.900488e+02, 2.866858e+02, 1.182969e+03, 2.751367e+02, 1.812147…
#> $ grain_size_suspended_ndim    <dbl> NA, NA, 700.88109, 730.32344, 101.02642, 897.67371, 1690.55153, 1293.03958, 2626.60876, 1266.72708, 3250.91379, 1749.93409, 3814…
#> $ grain_size_suspended_mm      <dbl> NA, NA, 0.059693451, 0.062201032, 0.008604335, 0.076454113, 0.143982849, 0.110127092, 0.223706055, 0.107886078, 0.276877588, 0.1…
#> $ mtpi30_min                   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ vb_bf_w_ratio                <dbl> NA, NA, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
#> $ da_scalar_maf                <dbl> NA, NA, 3.87597636, 3.86160608, 3.84921757, 3.82907072, 3.78997005, 3.76220325, 0.18256134, 0.17106844, 0.16309473, 0.15713507, …