Create Duration Suitability Rating Curve for a streamgage (scaled to prediction point by drainage area and precipitation)

Create Duration Suitability Rating Curve for a streamgage (scaled to prediction point by drainage area and precipitation)

habitat_drc_weighted(
  streamgage,
  reach,
  scale = "durhsi",
  habitat_type = "rearing",
  run = "fall",
  wy_group = "Dry",
  gradient = "Valley Foothill"
)

Arguments

streamgage

Three-letter CDEC identifier for the streamgage

reach

Integer comid identifier for reach identifying the prediction point (used for scaling)

scale

Either "durhsi" to keep original flow scale and adjust duration HSI values (default), or "flow" to keep original duration HSI scale and adjust flow values.

habitat_type

Either "rearing" (default) or "spawning"

run

One of "fall" (default), "late fall", "winter", "spring", or "steelhead"

wy_group

Either "Dry" or "Wet" for water year type scenario

gradient

For rearing, either "Valley Foothill" (default) or "Valley Lowland"

Value

A tibble with columns for flow (flow_cfs) and duration habitat suitability index (durhsi)

Examples

habitat_drc_weighted("FSB", habitat_type="rearing", gradient = "Valley Foothill", reach = 7978069)
#> # A tibble: 150 × 2
#>    flow_cfs durhsi
#>       <dbl>  <dbl>
#>  1      100   0.66
#>  2      200   0.66
#>  3      300   0.66
#>  4      400   0.66
#>  5      500   0.66
#>  6      600   0.66
#>  7      700   0.66
#>  8      800   0.66
#>  9      900   0.66
#> 10     1000   0.66
#> # ℹ 140 more rows