Predict Habitat Area using a Flow-to-Suitable-Area Curve

Predict Habitat Area using a Flow-to-Suitable-Area Curve

habitat_predict(
  x,
  reach,
  mainstem,
  watershed,
  streamgage,
  fsa_x,
  fsa_y,
  habitat_type = "rearing",
  units = "ft",
  run = "fall",
  wy_group = "Dry",
  gradient = "Valley Foothill"
)

Arguments

x

A vector of flows at which to predict suitable habitat values

reach

Integer comid identifier for reach (must provide either reach, mainstem, OR watershed)

mainstem

String name for mainstem (refer to cv_mainstems dataset for options)

watershed

String name for watershed (refer to cv_watersheds dataset for options)

streamgage

(optional) Three-letter CDEC identifier for a streamgage

fsa_x

(optional) Vector of flows for manually defined flow-to-suitable-area curve. If provided, overrides reach, mainstem, or watershed lookup.

fsa_y

(optional) Vector of habitat values for manually defined flow-to-suitable-area curve, of same length as fsa_x

habitat_type

Either "rearing" (default) or "spawning"

units

Desired units for output, either "ft" for square ft per linear ft, or "ac" for total acres

run

(required if streamgage is provided) One of "fall" (default), "late fall", "winter", "spring", or "steelhead"

wy_group

(required if streamgage is provided) Either "Dry" or "Wet" for water year type scenario

gradient

(required if streamgage is provided) For rearing, either "Valley Foothill" (default) or "Valley Lowland"

Value

A vector of the same length as x containing predicted suitable habitat values

Examples

# example non-temporal prediction given comid
habitat_predict(c(0, 10, 150, 300, 9999), reach = 7978069)
#> [1] 0.1450576 0.1450576 0.1576878 0.1605605 5.0666604

# example temporal prediction given comid and streamgage
habitat_predict(c(0, 10, 150, 300, 9999), reach = 7978069, streamgage = "FSB")
#> [1] 0.09784275 0.09784275 0.10407396 0.10596990 3.34399588

# example non-temporal prediction given mainstem
habitat_predict(c(0, 10, 150, 300, 9999), mainstem = "Feather River")
#> [1]  0.09182488  0.09182488  0.11423968  1.23489964 25.60762872

# example temporal prediction given comid and streamgage
habitat_predict(c(0, 10, 150, 300, 9999), mainstem = "Feather River", streamgage = "FSB")
#> [1]  0.06529095  0.06529095  0.07566989  0.81503042 17.19756125

# example manual prediction given custom provided flow-to-suitable-area curve
habitat_predict(c(0, 10, 150, 300, 9999), fsa_x = c(0, 100, 1000), fsa_y = c(47, 470, 4700))
#> [1]   47.000  258.500 1214.866 2488.223 4700.000