Logarithmic Sequence

Logarithmic Sequence

seq_log10(from, to, by = 0.1, snap)

Arguments

from

An integer defining the start of the sequence

to

An integer defining the end of the sequence

by

A number defining the exponent of 10 to increment by. Defaults to 0.1.

snap

An optional integer defining a midpoint of the sequence that should be preserved exactly in the range. If not defined, then from will be preserved exactly.

Value

An integer vector sequence

Examples

seq_log10(from = 50, to = 15000, by = 0.1, snap = 100)
#>  [1]    50.11872    63.09573    79.43282   100.00000   125.89254   158.48932
#>  [7]   199.52623   251.18864   316.22777   398.10717   501.18723   630.95734
#> [13]   794.32823  1000.00000  1258.92541  1584.89319  1995.26231  2511.88643
#> [19]  3162.27766  3981.07171  5011.87234  6309.57344  7943.28235 10000.00000
#> [25] 12589.25412